BREAKING CHANGE: Feature and Config rework add: xAllignment per Feature setting add: Mithril Grubber and Pure Quartz from Alpha server compat: ColorSelectionGui is now compatible with FeatureSetting fix: ButtonColorWheel not works fix: Highlight Enchant works while Parse Enchant Tooltips closed fix: SettingsGui cleanup fix: wrong dungeon teammate health fix: renderItem no longer clear depth buffer bit while currentScreen exist fix: Dragon Stats Tracker text mode setting not exist fix: minor fixes
Entire config management rewritten. With this, Feature
rework has been done. Testing was done individually, but
more testing would be good. Documentation will be added
in the future. I'm not very happy with this commit as it
includes too many changes, but since the config is spread
everywhere like a spider web, I decided to release it in
one piece. Finally, if the setting is off, the subsets
will now also appear as off. This can be changed in the
future.