add: parse dungeon players dungeon class level fix: Dungeon Healing Circle Wall doesn't work as expected in some cases fix: Dungeon Teammate Name Overlay's health line not affected by Text Style
The limitation here is that when there is more than
one healer in the team, the healer circle wall
solution will not work. The reason is that when
healers of different levels drop the ability on the
ground, the level of the owner of the circle on the
ground is unknown. It can be improved with more code
 in the future.