fix: even though player is in water, Pressure Text does not appear
Hypixel does not show the pressure value in the
action bar in case of mana consumption etc. In
this case, instead of not showing the Pressure
Text at all, we show the last remembered value
with the ‘~’ character. In the future, a mana
prediction style system could be used, with a
formula and pressure resistance value.